fix: provide 'empty' profile as fallback for fetching configs
