--- test.js
@@ -1,3 +1,0 @@
-import { join } from 'node:path'
-
-console.log(join('%root%', '.temp', 'shiny-config.json'))
